backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
geopy==2.4.0
geocoder==1.38.1
shapely==2.0.2
Pillow==10.0.1
boto3==1.34.0
stripe==7.6.0
twilio==8.10.0
sendgrid==6.10.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
alembic==1.12.1
